Initial commit- Made 2 stores into 1 store and removed mock data
