--- frontend/src/lib/store/index.ts
@@ -1,4 +1,0 @@
-// Re-export from the new location
-export * from '@/store/authStore';
-export * from '@/store/donationsStore';
-export * from '@/store/requestsStore';
